pyproject.toml
[project]
name = "cuga"
version = "0.2.10"
description = "CUGA is an open-source generalist agent for the enterprise, supporting complex task execution on web and APIs, OpenAPI/MCP integrations, composable architecture, reasoning modes, and policy-aware features."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10, <3.14"
dependencies = [
    "langchain>=1.0",
    "langchain-ibm>=0.3.9",
    "langchain-openai",
    "langchain-core",
    "aiohttp",
    "httpx",
    "docker",
    "langgraph",
    "markdownify",
    "uvicorn",
    "fastapi[standard]",
    "langfuse",
    "dynaconf",
    "playwright==1.49.0",
    "browsergym-core==0.13.0",
    "python-dotenv",
    "loguru",
    "mcp[cli]>=1.23.0",
    "psutil>=7.0.0",
    "typer>=0.15.3",
    "fastmcp",
    "langchain-mcp-adapters",
    "pyyaml>=6.0.3",
    "pymilvus[milvus-lite,model]>=2.6.4",
    "langchain-groq>=0.3.8",
    "langchain-litellm",
    "litellm>=1.0.0",
    "tavily-python",
    "torch>=2.2.2",
    "sentence-transformers>=5.1.2",
    "a2a-sdk[http-server]>=0.3.22",
    "sqlite-vec>=0.1.6",
    "pgvector>=0.4.2",
    "asyncpg>=0.30",
    "PyJWT[crypto]>=2.10.0",
    "cryptography>=44.0.0",
    "hvac>=2.0.0",
    "boto3>=1.34.0",
    "redis",
    "psycopg2-binary",
    "pydantic[email]>=2.0",
]

[project.optional-dependencies]
evaluation = ["rapidfuzz"]
e2b = ["e2b-code-interpreter>=2.4.1"]

memory = ["mem0ai[extras]"]

[project.urls]
Homepage = "https://cuga.dev"
Repository = "https://github.com/cuga-project/cuga-agent"

[tool.uv]
package = true

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["frontend_workspaces*"]

[tool.setuptools.package-data]
cuga = [
    "configurations/**/*.toml",
    "configurations/**/*.md",
    "settings.toml",
    "backend/**/*.jinja2",
    "backend/**/*.js",
]
frontend_workspaces = ["frontend/dist/**", "extension/releases/**"]
scripts = ["*.py"]
system_tests = ["**/*.py", "**/*.yaml", "**/*.sh"]


[project.scripts]
load_specs = "scripts.api_utils.load_openapi_specs:main"
petstore = "scripts.commands:run_petstore"
demo = "scripts.commands:run_demo"
registry = "scripts.commands:run_api_registry"
registry_appworld = "scripts.commands:run_api_registry_appworld"

graph = "scripts.draw_graph:main"
eval_api = "scripts.commands:run_eval_api"
appworld_eval = "evaluation.appworld_eval:run_main"
appworld_setup = "scripts.commands:setup_appworld_environment"
appworld_setup_docker = "scripts.commands:setup_appworld_environment_docker"
cuga = "cuga.cli:app"
code = "evaluation.code_generator:main"
exp = "scripts.eval_gui:main"
digital_sales_mcp = "scripts.commands:run_digital_sales_mcp"
digital_sales_openapi = "scripts.commands:run_digital_sales_openapi"
test_llm = "scripts.commands:test_llm"

[dependency-groups]
dev = [
    "pytz>=2025.1",
    "ruff>=0.9.4",
    "tabulate",
    "beartype",
    "nltk",
    "termcolor",
    "rapidfuzz",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=4.0",
    "cugaviz",
    "pre-commit>=4.3.0",
    "questionary>=2.1.1",
    "rich>=14.2.0",
    "pip-licenses>=5.5.1",
    "mypy>=1.0",
]
google-genai = ["langchain-google-genai>=2.0.11"]
sandbox = ["llm-sandbox>=0.1.3"]

[tool.pytest.ini_options]
anyio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
files = [
    "models.py",
    "merger.py",
    "db.py",
    "notifier_and_logger.py",
    "healthcheck.py",
    "outcome_tracker.py",
    "normalizers/",
]
